refactor: create user location hook
